2023-12-20 yeji commit1

+++ Krowd-POPing/src/main/java/com/jellybears/krowdpoping/report/model/dto/ReportProjectDTO.java
@@ -9,7 +9,7 @@ import lombok.*;
 @ToString
 public class ReportProjectDTO {
 
-    private int projectCode;
+    private Long projectCode;
     private String projectTitle;
 
 }
